Rename help panel question-related functionality
Indicate more clearly that (at least on the backend) the main
difference between the two methods of asking is not where the
question interface is shown but whether the question is posted
to the help desk or the mentor.

Also move the QuestionPoster classes into their own namespace,
in expectation of there being more classes soon, and make a
pseudo-constant into a real one.

There are no functional changes.

Bug: T250235
Change-Id: I543910bf174e0544b073df1acb11ad958f322ee2