Introduce MentorManager and virtually assign mentors to all users
Make mentor assignment happen on the fly, whenever the mentor
is first queried by any component. (Conceptually, this means
everyone is "assigned" a mentor when the patch is deployed, but
the identity of the mentor only gets stored once something
examines it. So the majority of users who never use the homepage
or the help panel won't have any mentor record, but code dealing
with mentors can just assume they always exist.)

This allows the "Ask your mentor" feature to be used in the help
panel, without having to worry about whether the given user has
a mentor assigned already.

To handle on-the-fly mentor assignment, and for better alignment
with coding standards, introduce MentorManager for setting and
retrieving mentors, turn Mentor into a value object and, move all
the business logic in Mentor into MentorPageMentorManager, and
update it to take advantage of available services. Some new methods
have a slightly different signature from their predecessor, but
there is only one important change: MentorManager::getMentorForUser
(the former Mentor::newFromMentee) does not have an $allowSelect
flag anymore, and always returns (and saves, if needed) a mentor.
The semantics of setting a mentor also changes slightly, to ensure
that setting a mentor takes effect immediately and there are no
race conditions.

It would make sense to rewrite or merge in ChangeMentor, but that's
left for a follow-up.

Bug: T250235
Bug: T259199
Change-Id: I3d4c89d7323d59ba75084a85e481ef3ab0b27936
Depends-On: Ie7d6e40d8387d8bc4f8592a31fdd70d0aad510ae