Fix help panel ask screen wording
Interface messages used to be defined in the HelpPanelProcessDialog
constructor and overridden via a bunch of configuration keys when
it was used on the homepage, but with the question target being
sort of independent from where the help panel is displayed (and
thus which part of the code it is invoked by) that's not practical
anymore, so merge the overrides back in and use a config flag to
select the right ones.

This is messy; in the future, moving the help panel screens into
separate classes and using inheritance to handle the mentor/helpdesk
variation would be a nicer approach.

Bug: T250235
Change-Id: Ibf5bc6c2bf45a03a5d9f9bab6fdf6c583afed35a