Add data for the top suggested edit card to the homepage
Fetch the top task suggestion for the given user and add it to the
page as a Javascript variable. (Actually using it for anything is
left to a separate patch.) To do that, replicate protection level
filtering on the server side. (Do not touch client-side filtering
as making the server-side version scale up to hundreds of tasks
would be a lot of effort, and hopefully the whole thing can be
replaced by a CirrusSearch filter eventually.)

As a side effect, UserSettingsDecorator is applied to the task
suggester by default now. This should have no effect to existing
functionality as the API always passes an array for the filter
parameters.

Also snuck in an unrelated fix to add
GrowthExperimentsServices::getQuestionPosterFactory which was
accidentally omitted in I2a59795c3.

Bug: T258021
Change-Id: I1f09f9f85357565865cb117ad07e44d80d9fe224