Fix "Ask mentor" help panel button styling
Swap styles for username and extra data, and fix spacing.

Also fix an unrelated issue with the icon positioning, which was
vertically centered in Zeplin but justified to bottom in practice;
and an unrelated linter warning.

The name "subsubheader" doesn't really make sense any more, but I
couldn't think of a better one.

Bug: T250235
Change-Id: I78d9315639b9f42233f668250515216e73992fb6