Replace UserSettingsDecorator with UserSettingsLookup
Replaces the decorator added in I6a9b05a1f with a convenience
helper for getting user settings. Using a decorator to change
the meaning of public interface parameters was a confusing use
of the pattern. Furthermore, the extra flexibility of having
the caller handle the option lookup allows for some performance
optimization.

Change-Id: I90dc04f4708e50081011f8718bd5e8355f84c549