Fix the first element of the suggested edits task queue on page load
When an initial task is provided by the server, display that immediately
(or as immediately as JS loading allows), and prepend it to the real
search results.

The result count will show "1" until the real results arrive, and
navigation will behave as if there would be only a single result.
Improving that is left to another commit.

Bug: T258024
Change-Id: I1a3fe023802876aabfc9dc38de7bccf05b947b6c