Add GrowthTasksApi.getDataFromActionApi
Loads the same data as fetchTasks, for a single task (except protection
info). Used for extending tasks exported from the server side.
We could do this on the server side but we need to load RESTBase/AQS
anyway so we might as well parallelize.

Includes some unrelated GrowthTasksApi refactoring: centralize (to the
extent possible) logging-related parameters, move some code to helpers,
make helpers consistently non-static and @private, get rid of the
comments on the task object in fetchTasks (already included in the
mw.libs.ge.TaskData typedef), fix a documentation typo.

Change-Id: I11478f66c34396b452133ad41deefb2c1709991a