Help panel: Reattach CTA button to overlay on startup
Currently, the button reanimates in some cases in guidance mode. This
happens when the infused server-side rendered button is relocated in the
DOM (attached to the overlay instead of #mw-content-text). This
reattachment is necessary to prevent the button from being hidden when
the editor is open, but it also causes reanimation.

Avoid reanimation by reattaching the button early, before it's even
visible. Fix the resulting strange-looking animation by applying the
guidance animation only to the help panel itself, not to the overlay
containing both the help panel and the button.

Bug: T255255
Change-Id: Ib83cc091ba28f3326c359b68d398d5e4d9dd143b