Remove header from discovery banner on mobile
If we are on the homepage, and the user is in variant C/D, don't show
the header text in the mobile discovery banner.

Also refactors SiteNoticeGenerator to accept an ExperimentUserManager in
the constructor, so we can add unit tests which check for different
output depending on the user's variant.

Bug: T258652
Change-Id: I909c25caae72fdb35fedb86455cd4d55eeae6127