Guidance: Also attach CTA button when opening help panel
We were correctly attaching the CTA button in guidance mode to account
for refreshing the page in read mode, but we were only doing this when
not trying to open the help panel. That was wrong, and caused the help
panel to open without the CTA button when refreshing.

Attach the CTA button regardless of whether we're about to open the help
panel.

Bug: T255255
Change-Id: I8e224a9c3dd6674b487e0136f3a5a4f476827c3e