Cache newcomer TaskSets per user
Implement a per-user cache for newcomer tasks, as encapsulated by the
TaskSet class.

When SearchTaskSuggester is invoked with either a local or remote
backend, look to see if we have cached a task set for the user who
made the request. If a task set is found, check to see if the task
and topic filters used to generate that task set (see TaskSetFilters)
are equal to the task and topic filters provided to SearchTaskSuggester.
If the filters match, then return that data to the user.

If no task set is found, or a task set is found but its filters do not
equal the filters provided to SearchTaskSuggester, then query the
search backend (local or remote) and store those results in the user's
cache.

Notes:

* The TTL for the cache is one week.
* When a TaskSet is generated (from cache or search), the suggestions
are shuffled. From a product perspective, the goal is for the user to
not see the same task card on page reload.
* The offset parameter is not handled in any way, since nothing was done
with it before the implementation of caching
* The debug parameter always bypasses the cache
* Add back the DecoratingTaskSuggesterFactory from I6a9b05a1f47069

For a follow-up:

* Implement a filter to ensure maintenance templates exist in result set
* Warm the cache for variant C users, who need to see a task card on
  first arrival to the homepage without first interacting with the
  task/topic filters

Bug: T260758
Change-Id: Ia5ea51860865b62f6700d3ad9491f38bc428bb17