Suggested Edits: Increase space between pager and card
Item B

This change brings the desktop and mobile implementations closer to
parity. The margin-bottom adds the equivalent amount of space as the
proposed padding-top: 16px to the card, except that our LESS file is
better set up to use margin-bttom on the pager (otherwise we'd need to
override the padding-top on the card for mobile).

Bug: T258704
Change-Id: Ia31964edc7759047b848b6d4b4ea01945834e732