Suggested Edits: Transparent border on activated module
Item G

Change the border color to transparent when the module is initiated on
desktop. The not-yet-initiated module keeps the existing @colorGray12
border color.

Note that the module can only be uninitiated in variant A and D; it's
initiated from the start in variant C. This code will also impact users
with variant A, who will see the transparent border.

Even though we are not targeting the "activated" class on mobile, it's
added in StartEditing dialog for consistency with desktop.

Bug: T258704
Change-Id: I1df67b452809f52daec7d54076173f2c6219def1