Always load the first task on the server side
For variant A, this will slow down rendering the page on the server side,
but will speed up rendering the suggested edits box on the client side.
The assumption is that suggested edits are the primary content of the
homepage so this is an OK tradeoff.

Reverts I7f3c7d94.

Bug: T258021
Change-Id: I64fcb8d4e8194390360e85a5d2d6bb8cc026d171