Show task preview in variant C/D mobile summary
(Mostly) reuse the styles from SmallTaskCard to show a task preview
card on the mobile summary of the suggested edits module. The card
is assembled on the server in a loading state, and lazy-loaded when
the JS module and the necessary API request finish. (Some day we'll
want to do this for the desktop homepage as well, with a different
HTML rendering, but that's not attempted in this patch.)

Remove GrowthTasksApi.getDataFromActionApi, which was unnecessary
complexity: we can just get the same data from PCS.

The nicer way to do this would probably be to create a PHP version
of the SmallTaskCard widget and use OOUI infusion, but I ran into
a number of issues when I tried that (using undefined for unloaded
state makes it impossible to correctly set the state in PHP, and
the way we load our modules does not seem compatible with infusion
either) which seemed to much effort to fix to be worth it.

Bug: T258021
Change-Id: I5ee77d29b96837b13ce9e2ba2fbe2b0303cbf376