Add TemplateFilter for validating TaskSet
Add a filter for checking that template based tasks in a cached TaskSet
are still valid. Each template based task corresponds to an article. The
task contains data about which maintenance templates are associated with
the task. Those templates should also be present on the article page.
Non cached TaskSets have a reasonably high chance of being accurate, as
the search index is updated quickly when maintenance templates are added
or removed. But for cached TaskSets, a task may no longer be valid
because the relevant maintenance template was removed from the article.

This patch applies a filter after fetching tasks in Suggested Edits
module and in ApiQueryGrowth tasks to double check that the relevant
templates still exist.

N.B. the TaskSet only contains a single task per article (per
SearchTaskSuggester::deduplicateSuggestions() ) so we don't have to
worry about edge cases of invalidating an item in the cache when another
valid task with a different set of templates (task type) exists.

For a follow-up:

* TemplateFilter and ProtectionFilter are pretty similar, perhaps they
should share some code
* TemplateFilterTest should use a data provider and cleanup in
between scenario setups and assertions

Bug: T260758
Change-Id: I74eb0a82f8c80937b36c2d7b9f7928c289f8827b