Limit caching to local tasks
Make TemplateFilter internal to CacheDecorator (as a side effect
this will apply template filtering to the task preview in the
mobile summary; it also makes the code a little simpler), and
only apply CacheDecorator to the local suggester.

This will have the downside of not testing the cache logic during
normal development, but we already have that issue with the split
code path for Local/RemoteTaskSuggester. OTOH not seeing cached
results during development is probably a plus.

The TemplateFilter service is preserved as the filter needs to be
injected via an ObjectFactory spec. That could be done without
a service but it would be uglier.

Also remove a duplicated call to ProtectionFilter.

Bug: T260758
Change-Id: Iac46a7ff83c7195c9506a420c03f0278735d0e9d