Impact: Clean up getTotalPageViews(), fix deprecation warning
When the Impact module displayed a page view count of 1000 or more, it
would throw a deprecation warning from Language::commafy() about
non-numeric input. This was because getTotalPageViews() returned a
number that was already formatNum()'ed (e.g. "1,442"), which was then
passed to ->numParams(), which attempts to format it again.

This wasn't the only thing that was wrong with getTotalPageViews(): its
output was used interchangeably as HTML or as a number, which meant that
it wasn't always being escaped properly.

Since the $maybeShowPendingIcon feature was only used by one caller
(getTotalViewsElement()), move it into that caller instead, and make
getTotalPageViews() simply return an integer.

Change-Id: Ie121bfaefef0557b907dcf4c08f239f421509e06