TaskTypeSelectionWidget: Use correct font size
This was falling back to the default font size for <h4>, which is
about right for modal dialogs on desktop (116% of 14px), but not
for dialogs that appear inside the content area on desktop (we don't
have these yet, but we will for variant D) or for dialogs on mobile.

Instead, set an explicit font size.

Change-Id: Ibcd75f589e12b72581ca9ebc9c0e654328dd076e