Use all task templates instead of current ones in TemplateFilter
This is conceptually more correct (replacing a task template
with an alias shouldn't invalidate the class), and reduces the
error surface - if ConfigurationLoader misbehaves, a number of
things will break anyway; this way we don't depends on
TemplateProvider as well.

Also rearrange the call order a bit to make it easier to
document, and drop a noop check.

Bug: T264029
Change-Id: I5ad9901894487e10456c195341db97bf5b1fd930