Properly handle namespaces in tasktype template configuration
Some wikis added the "Template:" prefix to template names, which is
an intuitive thing to do and the code should be able to handle it.
It ended up in the titletext part of TitleValue instead, causing
errors down the line. Fix it to parse template names properly.

Note that templates still have to be in the template namespace
(other parts of the code depend on that assumption, although
PageConfigurationLoader now doesn't anymore), but now they
should work with or without prefix.

Bug: T264029
Change-Id: Ibd3958b59cf75f31ccbc086382e20b6380733c3d
(cherry picked from commit ad13b0b598f61d18d47823dd05d2d61119148ab4)