SuggestedEdits: Fix footer display
Move the footer to the actual module footer, and fix the negative margin
rules to correct for the module div's padding instead of margin. Unwrap
the footer, and move the CSS rules to the actual footer div. Keep the
suggested-edits-footer class for the fake footer used in the mobile
summary.

In order to keep the footer always at the bottom, even when a max-height
rule makes the module taller than its natural height, use flexbox and
make the body take up the remaining height.

Also remove duplicative or frivolously different rules.

Bug: T258005
Bug: T258704
Change-Id: I0c044069dbf177d9fc766631c40a76f46aaa4dbd