Homepage: Render filter buttons on server-side
Render the filter buttons on the server-side, using the user's topic and
task filter preferences. The server-side functionality mostly follows
the client-side code in
SuggestedEditsFiltersWidget.prototype.updateButtonLabelAndIcon, any
differences are mainly due to the latter finding an existing button
group widget to work with (and having to handle the use case where a
user can unset all checkboxes).

Bug: T263040
Bug: T236738
Change-Id: I05fefaed9cf0c9a0b6dd2a374c7743a88340f87e