HomepageModule: Export taskcount, tasks and topics on server-side
If we have a cached taskset, export the task type, topics and task count
for logging as action_data.

To facilitate this, remove the limit parameter from suggest() when
obtaining the task set for server-side rendering of the card (otherwise
the task count is always limited to 10). Cache the count before it's
filtered down to a single task, and export the unfiltered count.

Bug: T262485
Change-Id: If51d4f1d8b33017e2c27421885a35b946edb5019