Convert node-qunit tests to qunit tests
Working with node-qunit was painful. While we lose some
features (running specific tests via the CLI, code coverage reports),
we gain some convenience in being able to set breakpoints in the browser
for interactive debugging, and our setup is more like the rest of the MW
ecosystem.

Overall I'm not thrilled about this change but also have the feeling it
will encourage us to write more JavaScript tests, which we haven't been
doing much of.

Also add a getEvents method to the logger classes since the tests are
run asynchronously, and we cannot subscribe to events pushed via
mw.track with any certainty about the contents of the queue.

Change-Id: I4c8c29cf9e04d7560cbbf3758c4f3f2154d85fb8