SE: Show no results card if we know there are no results
If we know there are no results, set a flag on the
server-side so that there is no immediate transition from attempting to
show a skeleton card to the no results card.

This was less noticeable in the pre-cache era, where the skeleton card
appeared for some moments before the API request to fetch tasks
completed.

Bug: T264314
Change-Id: Ib57478660333d892458552c92a33292b5088bf49