Use flex instead of breakpoints for homepage layout
Instead of breakpoints at set widths, which has limited
flexibility and interacted poorly with the new Vector skin
having a max-width for content, use a flexbox for the homepage
container, so that the main block and the sidebar take up
half-half of the excess space, and wrap when a minimum width of
480px and 300px, respectively, cannot be accommodated; and use
a flexbox for the two groups within the sidebar, with similar
rules (except 300/300 for min-width).

- Remove all variant C/D min-width, max-width and breakpoint-
  specific styles.
- Make the homepage container and the sidebar flex containers
  with wrapping in variant C/D.
- Make the main block, the sidebar and the sidebar subgroups
  flex boxes with a min width and growth and - except for the
  main block - shrink behavior.
- Fix homepage module class name generation to use
  growthexperiments-homepage-group-<group>-subgroup-<subgroup>-user-variant-<variant>
  consistent with all other classnames, instead of
  growthexperiments-homepage-<group>-subgroup-<subgroup>-user-variant-<variant>
  This should also fix the JS code from I84a456205 about mentor
  panel positioning which assumed consistent classnames.

Semi-related refactoring:
- Replace display: flex with the (slightly more compatible)
  .flex-display().
- Merge the last block of Homepage.less with the identical block
  that precedes it.
- Move some unrelated flex rules to the top of their block (not
  sure why CI didn't complain about them).

Note that new Vector has a min-width on .mw-page-container so flex
wrapping doesn't happen on small widths with that skin.

Bug: T264462
Change-Id: Ie0af5696e47dedf55bc326b9f1db995abf113d37