stylelint: Disable no-descending-specificity rule
We don't seem to care much about this rule in practice (it's disabled
on 13 lines already), it's hard to predict when it will happen,
it doesn't seem to add much value (I guess it was mainly meant for
CSS files where you can't rely on hierarchy to make base rules vs.
overrides somewhat self-documenting), it's usually a false positive
in the sense that it detects selector conflicts not rule conflicts
which is the actual problem it's supposed to avert, and stylelint's
habit of misreporting line numbers makes it particularly vexing to fix.

Change-Id: Idfb6680f624fdc2e84d444a4510fa701d3123417