Allow people who are not "fulltime mentors" to use Special:ClaimMentee
Course instructors often want to be a mentor of participants at their course,
but they do not want to be a mentor for everyone who comes.

This patch introduces a new page that contains people who can use
Special:ClaimMentee, despite not being listed at the main list of mentors.

Bug: T245639
Change-Id: I2323d7d462a78ce893f48594111aac7de0e69c7b