Newcomer tasks: Schedule a cache refresh via job queue
When CacheDecorator->suggest() is first called, the code calls out to
the search backend for retrieving results, and those results are cached
for one week. After the cache expires, the user will have a slow load
time when accessing Special:Homepage.

To work around this, whenever the cache is bypassed (the first time
we fetch tasks for a user; the cache is expired; or the cache is invalid
because the user has changed their filters), schedule a job to run a day
before the next cache expiration date. That job uses a new parameter to
suggest() to bypass retrieving cache results, but will save the new
results into the cache.

Bug: T260758
Depends-On: I9a9bc73e1094593fe4e2b72a8eae62569af1a42c
Change-Id: I958808389f8b131c17b639e8912669f74bc2a2cd