Suggested Edits: Server-side rendered card wrapper
Replace empty divs with a more complete construction of the HTML used on
the client-side. The only data we can render for the card so far is the
article title, but in a follow-up we can add the rendering for the task
explanation widget which will have the correct information for the task type.

This patch adds server-side renditions of the following JavaScript widgets:

* Edit card
* No results card
* Error card
* Previous / Next arrows

Bug: T263040
Bug: T236738
Change-Id: I81dd675376deda02a12c926e2b1679a47078508a