Do not fail with a generic error message "Something went wrong" when aborted with a non-good status
When the post question API returns an error code that isn't hookaborted,
like abusefilter-disallowed, display that error message, as it usually
has more information than our generic this.submitFailureMessage.

This should not affect ConfirmEdit (T266466), because that
extension doesn't fatal the status object, as it only needs
additional information to pass (CAPTCHA solution).

Bug: T266465
Change-Id: I4652bfede2ca34976991c0959d7cc817a4981bbe