Special:ClaimMentee should mention both lists in its error messages
This patch uses an if based on if mentorManager is an instance of
MentorPageMentorManager, and if it is, it tries to get the list of mentors.
If that succeeds, it mentions two lists, otherwise, it just mentions one.

If mentorManager is not MentorPageMentorManager, a generic message
is displayed, that does not mention any list. Currently, that will
only happen in local development while using StaticMentorManager.

Bug: T245639
Change-Id: Ib257f06d8aaca1bec925f01136237cfaec95253a