Fix NewcomerTasksCacheRefreshJob
While it is nowhere documented as far as I can see, jobs must have
a (title, params) constructor signature, unless they implement
GenericParameterJob, in which case it is just (params).
NewcomerTasksCacheRefreshJob used the latter signature, but did not
implement the interface, so it was passed a Title, and the params
got lost.

Bug: T268008
Change-Id: I9072243ce8ed170652062166fdf9d7b92d77b6d9