Introduce TaskTypeHandlerRegistry
Pulls together all the code which needs to branch on the task type.
Also makes it easy to add custom task types.

One tasktype-specific functionality is omitted: adding change tags
to edits. That requires changes to NewcomerTasks\Tracker so we know
the right task type; this is left to a separate patch in T266474.

Bug: T269366
Change-Id: I7d20b87c7391e84f6d40dba3f91fd3a24fa1676d
Depends-On: I787586fea2c3d9c3082e9d1007ecd2d552da6e4d