Remove tracking of individual templates from suggested edits
We tracked which specific maintenance templates a task's article
contained, and passed it to EventLogging, but never made use of the
data. With our focus now shifting to structured tasks, it is unlikely
we'd do it in the future. Remove template tracking to reduce code
complexity and increase performance.

The EventLogging schema field is set to empty, not removed, to avoid
changing the schema as it is being migrated.

Bug: T269366
Change-Id: If81f87e426eef72d26d5a3f75a146d7fefbb0e7e