Replace TemplateFilter with TaskSuggester::filter
Filtering out invalid tasks can be done by repeating the same Cirrus
query, with the set of pages fixed. This would require adding a
page: or pageid: filter to CirrusSearch (which is probably easy, as
these are already present in the index and name is already indexed
for prefix search), but would get rid of the more complex and less
performant TemplateFilter, and avoids the need for similar post-
filtering for category/template blocklists and for page protection,
and for recommendation validity.

Bug: T269642
Change-Id: Ib9c29a12f5cdb1f8d115590f1256e848a65397f6
Depends-On: Ic9eda738a3311ad781d26e65034f3e2fb35bcf4d