Unbreak quick start tips for link-recommendation for now
Add link-recommendation to TipLoader so that it doesn't throw an
exception, which causes the tips API endpoint to return an HTTP 500,
which crashes the frontend.

For some reason, reusing LinkTipTree for link-recommendation causes the
API endpoint to return an empty array. This causes the frontend to crash
every few seconds as it tries to auto-advance through an empty array, so
work around that as well.

Bug: T269654
Change-Id: Iad9d29e249510a790985b89c7244423b85ea951d