Allow use of link containers on the homepage for link click logging
Currently click targets that are supposed to emit a link-click event
must have a link-id data attribute. To support user-defined click
targets, introduce link-group-id, which is placed on some container
element. It is reported the same way; individual links within the
container are not differentiated.

Use it to report clicks to links inside the banner module.

Bug: T270309
Change-Id: I123b47e15793202bba8c28ca9a7b4699d49683e5