Track task type alongside with target page ID, step 1
Prepare for updating the tracking cache format so it can include
task type IDs. To make the change rollback-safe, in the first
step we don't actually change the format but make the code handle
the future format as well as the current one.

Also fix the issue where expiry was handled on a per-user level
so tasks never expired as long as the user kept visiting other
tasks.

Bug: T269366
Bug: T270639
Change-Id: Id322402d9ae4a56b1f8aece149a5c3bc5ab4f886