Add Link: Store user reviews of link recommendations
Create a new growthexperiments_link_submissions table which stores
what feedback users gave on individual recommended links. Negative
feedback is used to avoid recommending that link again. The data
can also be used to evaluate and improve the recommendation service,
and possibly for analytics (although for that purpose we'll probably
just duplicate the data into some more analytics oriented system).

Link recommendations are basically represented as task revid - link
pageid pairs. This is chosen to conserve storage, and because we
care more about the topic of the linked page than its name and want
to ignore renames.

Bug: T266446
Bug: T266473
Change-Id: I075bd1723bf297db49a54f052a3c8557cced79b4