Switch default help desk title to Main_Page
This is to make local development environment a little easier. By
default on a new wiki install, Talk:Main_Page won't have been created
and so the help panel won't load, and there won't be any log message to
explain why either. Setting to Main_Page allows the help panel to "just
work" after the preference is enabled.

This has no effect on production, since we override the value there for
all wikis we deploy to.

Change-Id: If7a09eb56c66649c5ea385134721bee89d00fe6d