EventGate: Fix missing module-name and mode for startediting-cta
To reproduce the problem, go to the mobile site as a variant C user and
click the "info" icon in the top right. The module name and mode are not
set because the .closest() method we rely on does not locate a homepage
module class. To work around this, embed the module name and mode for
variant C mobile context.

Bug: T273700
Change-Id: I0c86c713b24997077dfa0b439ec4f37fb1020081