Do not store the link recommendation in the DB on miss
Change the usage pattern of the growthexperiments_link_recommendation
table from a transparent cache layer to a persistence layer that needs
to be prefilled manually (already done by refreshLinkRecommendations.php).

Originally the service was assumed to provide recommendations, but due to
T266443 requiring filtering based on on-wiki parameters (which would be
difficult to handle within the service), those are more like recommendation
candidates, to be finalized by the refresh script, so we can't just directly
access the service. This would be especially problematic since we are going
to use the existence of a table row as a signal for the search index update.

Still allow falling back to the service (but without writing to the table)
via an optional configuration flag intended for debug environments, as
getting the refresh script to work in local environments is nontrivial
and whether the recommendation conforms exactly to the on-wiki rules is
probably irrelevant there.

Bug: T261409
Change-Id: I0037a43a278b9f33c2c5050d3c5a4833e39ed5cc