Reduce number of tasks requested and show full task count
Fetching 200 tasks at a time is slow. This isn't as much of problem for
loading Special:Homepage as the result set is cached, but it is slower
than it needs to be when adjusting task types and topics.

Instead, fetch 20 tasks at a time, and show the full task count instead
of the artificial limit of 200. If the user navigates to the end of the
20 item queue, fetch another 20 and filter out any duplicates. We don't
have an alternative for paging so this is the best we can do.

Modify the ApiQueryGrowthTasks endpoint to support accepting a set of
page IDs to exclude from the search query, to facilitate deduplicating
the task queue.

Also, update StaticTaskSuggester to use the title factory service, and
add filtering based on excluded page IDs for the selenium tests.

Bug: T260522
Change-Id: I4bde0105ebdc597084664a66d79b9877427da550