Rename ConfigurationLoader service to NewcomerTasks-specific name
I'm currently working on moving config from globals to a JSON
blob, that will be stored on-wiki. I would want to add a service
called "WikiPageConfigLoader" (or similar), which could be really
confusing with this newcomertasks-specific service.

I'll further refactor this to move PageLoader out of NewcomerTasks's
namespace in a separate commit.

Bug: T274522
Change-Id: I7416fe8e4ac078b9fdab265cd77112f088457457