Add WikiPageConfig that can provide config from wiki page
This is not currently used anywhere right now in the codebase,
it will be done in a follow-up patch, to make the patches cleaner.

The plan (for a next commit) is to get all config
loaded either through GrowthExperimentsGlobalVarConfig or
GrowthExperimentsWikiPageConfig, as decided in T275086.

GrowthExperimentsWikiPageConfig checks the allow list in
GEOnWikiConfigAllowList and then invokes WikiPageConfig. It also
fallbacks to GrowthExperimentsGlobalVarConfig, in case the on-wiki
page does not have the config. This prevents wikis totally breaking
the extension by _removing_ a config variable, and also makes developer
setups work without any config page created locally.

GrowthExperimentsGlobalVarConfig is a different name for
GlobalVarConfig.

Configuration variables mapping is taken from T275086.

Bug: T274520
Change-Id: I32145d2538dd213171f768e42116c2545f00321b