Make Special:WelcomeSurvey work without group parameter when directly viewed
It is useful for communities trying out our features to be
able to try welcome survey easily.

Before this commit, it just redirects away when no experiment
is active at that wiki. It can be overridden by _group query parameter,
but that is not widely known. It's more user friendly if directly entering
Special:WelcomeSurvey would just work.

This commit accomplishes that by introducing a new variable, "default variant",
that is used when the user directly visits Special:WelcomeSurvey.

The WelcomeSurveyHooks::onBeforeWelcomeCreation hook sets the group
into user preference before redirecting to the survey,
so that should not be affected.

Bug: T274368
Change-Id: I8f8ecccada9a87f05d20e19bcc2b613820c6645e