Make it possible to disable mentorship features
We want to have a way to disable mentorship features
for everyone, not just for new users by setting
the mentors list to null.

Add a new global variable that can do this. It's not clear
whether communities should be allowed to alter it when
config is moved to a wiki JSON, but that will be disucssed
and decided later as part of T274520.

This affects both the homepage mentorship module
and help panel mentorship features.

Bug: T273776
Change-Id: Ia85f4a2872f2a02fab8379ea7f0b3c522b210384