Impact module: Add "not rendered" state
impact_module_state is a required field, so when a wiki doesn't have
this module, we still need to record a state (unless we want to create a
new major version of the schema which didn't seem like the better option
here). So, add a NOTRENDERED constant for capturing this cenario and use
it in Impact module's getState().

Postscript: Because only testwiki is using secondary/jsonschema, and
since the HomepageVisit page is locked from editing on metawiki, use
"unactivated" as a placeholder for not rendered.

Bug: T270294
Bug: T275615
Change-Id: I0c35bc7a49e0a074e5bc835dae88f17efe9e3f2b
(cherry picked from commit 984c378507ce7093be45a59d4eb5aca3ebeb83a4)