Help panel: Do not require help desk to be configured
As part of scaling Growth features across multiple wikis,
we want to be able to deploy to wikis with no local help desk.

Soon, we are going to switch to mentorship by default,
but help desk is still going to be used as a fallback. If even
help desk is not around, no ask-help module should be displayed.

Bug: T273118
Change-Id: I6bf58ca26d29a1badc812bd48a6902a4b4cfad92