Add fixLinkRecommendationData.php maintenance script
Add a maintenance script for ensuring that the recommendations table
and the search index is in sync. This is useful for fixing test setups
if things get messed up somehow.

No attempt is made to handle replication lag, delayed search index
updates due to job queue size or batching, and similar potential race
conditions. As such, the script is not appropriate for production use.

Bug: T274198
Change-Id: I861db8702b30c4559a19c1bdd6d108d12389fc23