Make GrowthExperimentsMultiConfig::get work when no config page exists
Right now, GrowthExperimentsMultiConfig::get unconditionally throws
if no config page is found, because WikiPageConfig::getConfigData
throws in case it fails to find the config page. Instead, make getConfigData
return an empty array in this scenario, so it behaves in the same way
as if the page was an empty dictionary.

This is to make it possible to use the new config logic without actually
touching anything on-wiki.

Bug: T274520
Change-Id: Ic5ae39d2625fef26ae672742eb8350ca72567f8d