Filter link recommendation links on demand
Filter out links which have been rejected more than 2 times.
We want rejections to take effect reasonably quickly, so we
must do this whenever we load a cached recommendation.
Also filter out redlinks (this happens when a page got
deleted since the last database update of the recommendation
service; such links look ugly as we can't show any details,
and there is a high chance they are incorrect).

When the recommendation is left with no valid links at all,
suppress it.

To avoid duplication of code with refreshLinkRecommendations,
put it into AddLinkHandlerTrait (which already wasn't really
specific to REST handlers), and turn it into a helper class
(as it needs more dependencies which is awkward for a trait).

Also use a custom exception class instead of the RestHandler-
specific HttpException.

Bug: T266437
Change-Id: I5749c460bcbef454346769941072b94442ec8431