Guard against undefined suggestion
While iterating over suggestions from the phrase map, we don't have a
guarantee that the index corresponding to occurences seen will be
populated with a suggestion.

This particular bug surfaces when the phrase to be linked occurs more
than once in a document, e.g. "Berlin" might appear in the first and
second paragraphs and only one of the matches is to be proposed as a
link.

Bug: T276838
Change-Id: I2549dba722dd148549545582e21255ff3ea2cbe5