Delete link recommendation from DB when it becomes invalid
Delete link recommendation from the database when it is no longer
valid: when some of the recommended links were rejected, when some
of the recommended links were accepted and when a non-suggested
edit was made to the article. The first is a bugfix (the DB record
staying around would prevent the article from selected as a task
again), the other two are cleanups with no functional change.

Bug: T275790
Change-Id: I5d74aee80f84c2317547035c4196635586231755