Add WikiPageConfigValidation
This class serves as the canonical place to store
form descriptor for the config, will be used in the
special page as well.

The default value is stored in validator class. We can
move it somewhere else in a future commit if we want,
but for now, it makes sense to do it here, IMO.

Bug: T274031
Change-Id: I72e400fd2344589140a99d0890b53d917086a5b1