Monitoring: Add endpoint for metadata about suggested edit task pool
Return an array of data with keys for tasks, topics and totalCount.

Tasks is an array containing the task type ID and the count available
for each task.

Topics is an array containing a list of tasks, and the count for the
number of tasks available for each task type and topic. A total count
for the topic is provided as well.

Bug: T276795
Bug: T278413
Bug: T249987
Change-Id: I095163fecf9f438b52d34d5341ad9b212c42d829