Add listTaskCounts.php option to report counts to statsd or as JSON
Add --statsd option to report counts to statsd, with keys names like
MediaWiki.<wiki-id>.growthexperiments.taskcount.<task-type-id>.<topic-id>
and MediaWiki.<wiki-id>.growthexperiments.tasktypecount.<task-type-id>
(the number of tasks for a given task type in a given topic and the
number of tasks for a given task type regardless of topic, respectively).

Currently only done for the link recommendation task type to keep the
metric count down as we don't have an alerting use case for the other
metrics.

Also add a JSON output so one can use jq to conveniently calculate
totals etc.

Bug: T278411
Change-Id: I9f30a60ecfb96e20a6ba66c70cca1384faa0dee6